HARP-3994: Move outlines to solid outlines + GeoJsonDecoder refactoring.
