+++ @here/harp-geojson-datasource/lib/GeoJsonDecoder.ts
@@ -8,149 +8,25 @@ import {
     DecodedTile,
     ExtendedTileInfo,
     ExtendedTileInfoWriter,
-    FillTechnique,
-    Geometry,
-    GeometryType,
-    getPropertyValue,
-    Group,
-    IMeshBuffers,
-    InterleavedBufferAttribute,
-    isCirclesTechnique,
-    isFillTechnique,
-    isPoiTechnique,
-    isSolidLineTechnique,
-    isSquaresTechnique,
-    isTextTechnique,
-    LineFeatureGroup,
-    Outliner,
-    PoiGeometry,
-    StyleSetEvaluator,
-    Technique,
-    TextGeometry,
-    TextPathGeometry
+    StyleSetEvaluator
 } from "@here/harp-datasource-protocol";
-import { MapEnv, Value } from "@here/harp-datasource-protocol/lib/Theme";
-import { GeoCoordinates, Projection, TileKey, webMercatorTilingScheme } from "@here/harp-geoutils";
-import { LINE_VERTEX_ATTRIBUTE_DESCRIPTORS, Lines } from "@here/harp-lines";
+import { Projection, TileKey, webMercatorTilingScheme } from "@here/harp-geoutils";
 import { ThemedTileDecoder } from "@here/harp-mapview-decoder";
 import { WorkerServiceManager } from "@here/harp-mapview-decoder/index-worker";
 import { TileDecoderService } from "@here/harp-mapview-decoder/lib/TileDecoderService";
-import { LoggerManager, Math2D } from "@here/harp-utils";
-import earcut from "earcut";
+import { LoggerManager } from "@here/harp-utils";
 import * as THREE from "three";
-import { Feature, FeatureDetails, GeoJsonDataType } from "./GeoJsonDataType";
-import { Flattener } from "./utils/Flattener";
+import { GeoJson } from "./GeoJsonDataType";
+import { GeoJsonGeometryCreator, GeoJsonTileGeometries } from "./GeoJsonGeometryCreator";
+import { ExtendedTile } from "./GeoJsonParser";
 
 const logger = LoggerManager.instance.create("GeoJsonDecoder");
-
-/**
- * Store temporary variables needed for the creation of a mesh.
- */
-class MeshBuffer implements IMeshBuffers {
-    readonly positions: number[] = [];
-    readonly indices: number[] = [];
-    readonly groups: Group[] = [];
-    readonly outlineIndices: number[][] = [];
-    readonly edgeIndices: number[] = [];
-    /**
-     * Optional list of feature start indices. The indices point into the index attribute.
-     */
-    readonly featureStarts: number[] = [];
-    /**
-     * Optional list of feature IDs. Currently only Number is supported, will fail if features have
-     * IDs with type Long.
-     */
-    readonly featureIds: Array<number | undefined> = [];
-    /**
-     * Optional object containing the geojson properties defined by the end-user.
-     */
-    readonly geojsonProperties: Array<{} | undefined> = [];
-}
-
-/**
- * Store temporary variables needed for the creation of a geometry.
- */
-class GeometryData {
-    /**
-     * XYZ defines the property to label in the style, which we cannot do. Since all the features
-     * of a same type will show the same property in a given datasource, this workaround is to
-     * hand the name property to pick to the buffer.
-     */
-    labelProperty?: string;
-    type: string | undefined;
-    points: PointsData = { vertices: [], geojsonProperties: [undefined] };
-    lines: LinesData = { vertices: [], geojsonProperties: [] };
-    polygons: PolygonsData[] = [];
-}
-
-/**
- * Interface used to store vertices, holes and properties of polygons.
- */
-interface PolygonsData {
-    vertices: number[];
-    holes: number[];
-    geojsonProperties?: {};
-}
-
-/**
- * Interface used to store vertices and properties of points.
- */
-interface PointsData {
-    vertices: number[];
-    geojsonProperties: [{} | undefined];
-}
-
-/**
- * It stores the vertices of each line. The lenght of the property `vertices` and that one of
- * `geojsonProperties` should be the same.
- */
-interface LinesData {
-    vertices: number[][];
-    geojsonProperties: Array<{} | undefined>;
-}
-
-/**
- * Interface use to read and write from/to an [[ExtendedTile]].
- */
-interface ExtendedTile {
-    info: ExtendedTileInfo;
-    writer: ExtendedTileInfoWriter;
-}
 
 /**
  * Interface for the GeoJsonDecodedTile.
  */
 export interface GeoJsonDecodedTile extends DecodedTile {
     tileMortonCode: number;
-}
-
-/**
- * Geometry interface that stores the geometry type and the user data.
- */
-export interface GeoJsonGeometry extends Geometry {
-    type: GeometryType;
-    objInfos?: Array<{} | undefined>;
-}
-
-/**
- * Geometry interface that stores the geometry type and the user data for a POI.
- */
-export interface GeoJsonPoiGeometry extends PoiGeometry {
-    objInfos?: Array<{} | undefined>;
-}
-
-/**
- * Geometry interface that stores the geometry type and the user data for a Text.
- */
-export interface GeoJsonTextGeometry extends TextGeometry {
-    objInfos?: Array<{} | undefined>;
-}
-
-/**
- * Geometry interface that stores the geometry type and the user data for a TextPath.
- */
-export interface GeoJsonTextPathGeometry extends TextPathGeometry {
-    objInfos?: Array<{} | undefined>;
 }
 
 /**
@@ -199,7 +75,7 @@ export class GeoJsonTileDecoder extends ThemedTileDecoder {
      * @param projection The current camera projection.
      */
     decodeThemedTile(
-        data: GeoJsonDataType,
+        data: GeoJson,
         tileKey: TileKey,
         styleSetEvaluator: StyleSetEvaluator,
         projection: Projection
@@ -212,25 +88,10 @@ export class GeoJsonTileDecoder extends ThemedTileDecoder {
 }
 
 /**
- * Decoder for the GeoJson datasource.
+ * Decoder for the [[GeoJsonDataSource]].
  */
 class GeoJsonDecoder {
-    private m_center: THREE.Vector3 = new THREE.Vector3();
-    private readonly m_geometries: GeoJsonGeometry[] = [];
-    private m_textGeometries?: GeoJsonTextGeometry[];
-    private m_textPathGeometries?: TextPathGeometry[];
-    private m_poiGeometries?: PoiGeometry[];
-
-    private readonly m_geometryBuffer = new Map<number, GeometryData>();
-    private readonly m_textGeometryBuffer = new Map<number, GeometryData>();
-    private readonly m_textPathGeometryBuffer = new Map<number, GeometryData>();
-    private readonly m_poiGeometryBuffer = new Map<number, GeometryData>();
-
     private readonly m_storeExtendedTags = true;
-    private readonly m_gatherRoadSegments = true;
-    private m_cached_worldCoord: THREE.Vector3 = new THREE.Vector3();
-    private m_cached_geoCoord: GeoCoordinates = new GeoCoordinates(0, 0);
-    private m_outliner: Outliner = new Outliner();
 
     /**
      * Default constructor.
@@ -248,8 +109,8 @@ class GeoJsonDecoder {
      *
      * @returns decoded tile
      */
-    getDecodedTile(tileKey: TileKey, data: GeoJsonDataType): DecodedTile {
-        this.m_center = this.getTileCenter(tileKey);
+    getDecodedTile(tileKey: TileKey, data: GeoJson): DecodedTile {
+        const center = this.getTileCenter(tileKey);
         const tileInfo = new ExtendedTileInfo(tileKey, this.m_storeExtendedTags);
         const tileInfoWriter = new ExtendedTileInfoWriter(tileInfo, this.m_storeExtendedTags);
 
@@ -257,991 +118,34 @@ class GeoJsonDecoder {
             info: tileInfo,
             writer: tileInfoWriter
         };
-        this.processGeoJson(data, extendedTile);
+
+        const geometries: GeoJsonTileGeometries = GeoJsonGeometryCreator.createGeometries(
+            data,
+            center,
+            this.m_projection,
+            this.m_styleSetEvaluator,
+            extendedTile
+        );
 
         const tile: DecodedTile = {
-            geometries: this.m_geometries,
+            geometries: geometries.geometries,
             techniques: this.m_styleSetEvaluator.techniques,
             tileInfo: this.getTileInfo(extendedTile)
         };
 
-        if (this.m_poiGeometries !== undefined) {
-            tile.poiGeometries = this.m_poiGeometries;
+        if (geometries.poiGeometries.length > 0) {
+            tile.poiGeometries = geometries.poiGeometries;
         }
 
-        if (this.m_textGeometries !== undefined) {
-            tile.textGeometries = this.m_textGeometries;
+        if (geometries.textGeometries.length > 0) {
+            tile.textGeometries = geometries.textGeometries;
         }
 
-        if (this.m_textPathGeometries !== undefined) {
-            tile.textPathGeometries = this.m_textPathGeometries;
+        if (geometries.textPathGeometries.length > 0) {
+            tile.textPathGeometries = geometries.textPathGeometries;
         }
 
         return tile;
-    }
-
-    /**
-     * Processes the GeoJSON.
-     *
-     * @param data source GeoJSON data
-     */
-    private processGeoJson(data: GeoJsonDataType, extendedTile: ExtendedTile): void {
-        switch (data.type) {
-            case "Feature":
-                this.processFeature(data, extendedTile);
-                break;
-            case "FeatureCollection":
-                data.features.forEach(feature => this.processFeature(feature, extendedTile));
-                break;
-            default:
-                const collectionFeature: Feature = {
-                    type: "Feature",
-                    geometry: data
-                };
-                this.processFeature(collectionFeature, extendedTile);
-                break;
-        }
-
-        this.createGeometry(extendedTile.info);
-    }
-
-    /**
-     * Processes GeoJSON "Feature" and "GeometryCollection" objects.
-     *
-     * @param feature source Feature data
-     */
-    private processFeature(feature: Feature, extendedTile: ExtendedTile): void {
-        // Skip features without geometries.
-        if (feature.geometry === null) {
-            return;
-        }
-
-        let techniqueIndices: number[] = [];
-        switch (feature.geometry.type) {
-            case "Point":
-                techniqueIndices = this.findTechniqueIndices(feature, "point");
-                for (const techniqueIndex of techniqueIndices) {
-                    const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-                    if (isPoiTechnique(technique)) {
-                        this.processPoi(
-                            [feature.geometry.coordinates],
-                            techniqueIndex,
-                            feature.properties
-                        );
-                    } else if (isTextTechnique(technique)) {
-                        this.processText(
-                            [feature.geometry.coordinates],
-                            techniqueIndex,
-                            technique.label!,
-                            feature.properties
-                        );
-                    } else if (isCirclesTechnique(technique) || isSquaresTechnique(technique)) {
-                        this.processPoint(
-                            [feature.geometry.coordinates],
-                            techniqueIndex,
-                            feature.properties
-                        );
-                    }
-                }
-                break;
-
-            case "MultiPoint":
-                techniqueIndices = this.findTechniqueIndices(feature, "point");
-                for (const techniqueIndex of techniqueIndices) {
-                    const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-                    if (isPoiTechnique(technique)) {
-                        this.processPoi(
-                            feature.geometry.coordinates,
-                            techniqueIndex,
-                            feature.properties
-                        );
-                    } else if (isTextTechnique(technique)) {
-                        this.processText(
-                            feature.geometry.coordinates,
-                            techniqueIndex,
-                            technique.label!,
-                            feature.properties
-                        );
-                    } else if (isCirclesTechnique(technique) || isSquaresTechnique(technique)) {
-                        this.processPoint(
-                            feature.geometry.coordinates,
-                            techniqueIndex,
-                            feature.properties
-                        );
-                    }
-                }
-                break;
-
-            case "LineString":
-                techniqueIndices = this.findTechniqueIndices(feature, "line");
-                for (const techniqueIndex of techniqueIndices) {
-                    const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-                    if (isSolidLineTechnique(technique)) {
-                        this.processLineString(
-                            extendedTile,
-                            [feature.geometry.coordinates],
-                            techniqueIndex,
-                            feature.id,
-                            feature.properties
-                        );
-                    } else if (isTextTechnique(technique)) {
-                        this.processLineText(
-                            feature.geometry.coordinates,
-                            techniqueIndex,
-                            technique.label!,
-                            feature.properties
-                        );
-                    }
-                }
-                break;
-
-            case "MultiLineString":
-                techniqueIndices = this.findTechniqueIndices(feature, "line");
-                for (const techniqueIndex of techniqueIndices) {
-                    const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-                    if (isSolidLineTechnique(technique)) {
-                        this.processLineString(
-                            extendedTile,
-                            feature.geometry.coordinates,
-                            techniqueIndex,
-                            feature.id,
-                            feature.properties
-                        );
-                    } else if (isTextTechnique(technique)) {
-                        for (const coordinate of feature.geometry.coordinates) {
-                            this.processLineText(
-                                coordinate,
-                                techniqueIndex,
-                                technique.label!,
-                                feature.properties
-                            );
-                        }
-                    }
-                }
-                break;
-
-            case "Polygon":
-                techniqueIndices = this.findTechniqueIndices(feature, "polygon");
-                for (const techniqueIndex of techniqueIndices) {
-                    const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-                    if (isFillTechnique(technique)) {
-                        this.processPolygon(
-                            [feature.geometry.coordinates],
-                            techniqueIndex,
-                            feature.properties
-                        );
-                    } else if (isTextTechnique(technique)) {
-                        for (const coordinate of feature.geometry.coordinates) {
-                            this.processPolygonLabel(
-                                coordinate,
-                                techniqueIndex,
-                                technique.label!,
-                                feature.properties
-                            );
-                        }
-                    }
-                }
-                break;
-
-            case "MultiPolygon":
-                techniqueIndices = this.findTechniqueIndices(feature, "polygon");
-                for (const techniqueIndex of techniqueIndices) {
-                    const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-                    if (isFillTechnique(technique)) {
-                        this.processPolygon(
-                            feature.geometry.coordinates,
-                            techniqueIndex,
-                            feature.properties
-                        );
-                    } else if (isTextTechnique(technique)) {
-                        for (const polygons of feature.geometry.coordinates) {
-                            for (const polygon of polygons) {
-                                this.processPolygonLabel(
-                                    polygon,
-                                    techniqueIndex,
-                                    technique.label!,
-                                    feature.properties
-                                );
-                            }
-                        }
-                    }
-                }
-                break;
-
-            case "GeometryCollection":
-                feature.geometry.geometries.forEach(geometry => {
-                    const collectionFeature: Feature = {
-                        type: "Feature",
-                        properties: feature.properties,
-                        geometry
-                    };
-                    this.processFeature(collectionFeature, extendedTile);
-                });
-                break;
-
-            default:
-                logger.warn("Invalid GeoJSON data. Unknown geometry type.");
-        }
-    }
-
-    /**
-     * Creates a point at the center of each line segment for lines that have a text technique.
-     *
-     * @param coordinates Array containing coordinates.
-     * @param techniqueIndex Technique index to render geometry.
-     * @param geojsonProperties Object containing the properties defined by the user.
-     */
-    private processLineText(
-        coordinates: number[][],
-        techniqueIndex: number,
-        labelProperty: string,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreateTextPathGeometryBuffer(techniqueIndex);
-        buffer.type = "text-path";
-        buffer.labelProperty = labelProperty;
-
-        const vertices: number[] = [];
-
-        for (const point of coordinates) {
-            this.m_cached_geoCoord.latitude = point[1];
-            this.m_cached_geoCoord.longitude = point[0];
-            this.m_projection
-                .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                .sub(this.m_center);
-
-            vertices.push(this.m_cached_worldCoord.x, this.m_cached_worldCoord.y);
-        }
-        buffer.lines.vertices.push(vertices);
-        buffer.lines.geojsonProperties.push(geojsonProperties);
-    }
-
-    /**
-     * Creates a point at the center of a polygon feature that matches a text technique.
-     *
-     * @param coordinates Array containing coordinates.
-     * @param techniqueIndex Technique index to render geometry.
-     * @param geojsonProperties Object containing the properties defined by the user.
-     */
-    private processPolygonLabel(
-        coordinates: number[][],
-        techniqueIndex: number,
-        labelProperty: string,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreateTextGeometryBuffer(techniqueIndex);
-        buffer.type = "point";
-        buffer.labelProperty = labelProperty;
-
-        const points = { x: [] as number[], y: [] as number[], z: [] as number[] };
-
-        coordinates.forEach(point => {
-            this.m_cached_geoCoord.latitude = point[1];
-            this.m_cached_geoCoord.longitude = point[0];
-
-            this.m_projection
-                .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                .sub(this.m_center);
-
-            points.x.push(this.m_cached_worldCoord.x);
-            points.y.push(this.m_cached_worldCoord.y);
-            points.z.push(this.m_cached_worldCoord.z);
-        });
-
-        this.m_cached_worldCoord.setX(points.x.reduce((a, b) => a + b) / coordinates.length);
-        this.m_cached_worldCoord.setY(points.y.reduce((a, b) => a + b) / coordinates.length);
-        this.m_cached_worldCoord.setZ(points.z.reduce((a, b) => a + b) / coordinates.length);
-
-        buffer.points.vertices.push(
-            this.m_cached_worldCoord.x,
-            this.m_cached_worldCoord.y,
-            this.m_cached_worldCoord.z
-        );
-        const pointIndex = buffer.points.vertices.length / 3 - 1;
-        buffer.points.geojsonProperties[pointIndex] = geojsonProperties;
-    }
-
-    /**
-     * Processes the GeoJSON's "Point" and "MultiPoint" objects when a POI technique is provided.
-     *
-     * @param coordinates Array containing coordinates.
-     * @param techniqueIndex Technique index to render geometry.
-     * @param geojsonProperties Object containing the properties defined by the user.
-     */
-    private processPoi(
-        coordinates: number[][],
-        techniqueIndex: number,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreatePoiGeometryBuffer(techniqueIndex);
-        buffer.type = "poi";
-
-        coordinates.forEach(point => {
-            this.m_cached_geoCoord.latitude = point[1];
-            this.m_cached_geoCoord.longitude = point[0];
-
-            this.m_projection
-                .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                .sub(this.m_center);
-            buffer.points.vertices.push(
-                this.m_cached_worldCoord.x,
-                this.m_cached_worldCoord.y,
-                this.m_cached_worldCoord.z
-            );
-            const pointIndex = buffer.points.vertices.length / 3 - 1;
-            buffer.points.geojsonProperties[pointIndex] = geojsonProperties;
-        });
-    }
-
-    /**
-     * Processes the GeoJSON objects when a Text technique is provided.
-     *
-     * @param coordinates Array containing coordinates.
-     * @param techniqueIndex Technique index to render geometry.
-     * @param geojsonProperties Object containing the properties defined by the user.
-     */
-    private processText(
-        coordinates: number[][],
-        techniqueIndex: number,
-        labelProperty: string,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreateTextGeometryBuffer(techniqueIndex);
-        buffer.type = "text";
-        buffer.labelProperty = labelProperty;
-
-        coordinates.forEach(point => {
-            this.m_cached_geoCoord.latitude = point[1];
-            this.m_cached_geoCoord.longitude = point[0];
-
-            this.m_projection
-                .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                .sub(this.m_center);
-            buffer.points.vertices.push(
-                this.m_cached_worldCoord.x,
-                this.m_cached_worldCoord.y,
-                this.m_cached_worldCoord.z
-            );
-            const pointIndex = buffer.points.vertices.length / 3 - 1;
-            buffer.points.geojsonProperties[pointIndex] = geojsonProperties;
-        });
-    }
-
-    /**
-     * Processes the GeoJSON's "Point" and "MultiPoint" objects.
-     *
-     * @param coordinates Array containing coordinates.
-     * @param techniqueIndex Technique index to render geometry.
-     * @param geojsonProperties Object containing the properties defined by the user.
-     */
-    private processPoint(
-        coordinates: number[][],
-        techniqueIndex: number,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreateGeometryBuffer(techniqueIndex);
-        buffer.type = "point";
-
-        coordinates.forEach(point => {
-            this.m_cached_geoCoord.latitude = point[1];
-            this.m_cached_geoCoord.longitude = point[0];
-
-            this.m_projection
-                .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                .sub(this.m_center);
-            buffer.points.vertices.push(
-                this.m_cached_worldCoord.x,
-                this.m_cached_worldCoord.y,
-                this.m_cached_worldCoord.z
-            );
-            const pointIndex = buffer.points.vertices.length / 3 - 1;
-            buffer.points.geojsonProperties[pointIndex] = geojsonProperties;
-        });
-    }
-
-    /**
-     * Processes the GeoJSON's "LineString" and "MultiLineString" objects.
-     *
-     * @param coordinates Array containing line coordinates.
-     * @param techniqueIndex Technique index to use for the geometry rendering.
-     * @param geojsonProperties Object containing the custom properties defined by the user
-     */
-    private processLineString(
-        extendedTile: ExtendedTile,
-        coordinates: number[][][],
-        techniqueIndex: number,
-        featureId?: string,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreateGeometryBuffer(techniqueIndex);
-        buffer.type = "solid-line";
-
-        coordinates.forEach(line => {
-            buffer.lines.geojsonProperties.push(geojsonProperties);
-            const vertices: number[] = [];
-            line.forEach(point => {
-                if (point === null || point[0] === null || point[0] === undefined) {
-                    return;
-                }
-                this.m_cached_geoCoord.latitude = point[1];
-                this.m_cached_geoCoord.longitude = point[0];
-                this.m_projection
-                    .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                    .sub(this.m_center);
-                vertices.push(this.m_cached_worldCoord.x, this.m_cached_worldCoord.y);
-            });
-
-            buffer.lines.vertices.push(vertices);
-        });
-
-        const featureDetails: FeatureDetails = {};
-        if (featureId !== undefined) {
-            featureDetails.featureId = featureId;
-        }
-
-        const env = new MapEnv({ type: "line", ...featureDetails });
-        const techniques = this.m_styleSetEvaluator.getMatchingTechniques(env);
-        const featureIdNumber = 0; //geojsonTile do not have an integer for the featureId. Use 0.
-        if (buffer.lines.vertices.length !== buffer.lines.geojsonProperties.length) {
-            logger.log("The amount of lines and the amount of geo properties has to be the same");
-            return;
-        }
-        this.addTileInfo(
-            extendedTile,
-            techniques,
-            buffer.lines.vertices,
-            featureIdNumber,
-            env,
-            buffer.lines.geojsonProperties
-        );
-    }
-
-    private addTileInfo(
-        extendedTile: ExtendedTile,
-        techniques: Technique[],
-        lines: number[][],
-        featureId: number,
-        env: MapEnv,
-        geojsonProperties: Array<{} | undefined>
-    ) {
-        if (geojsonProperties.length !== lines.length) {
-            logger.error("geojsonProperties and lines should have the same lenght");
-            return;
-        }
-        const tileInfoWriter = extendedTile.writer;
-        for (const technique of techniques) {
-            if (technique === undefined) {
-                continue;
-            }
-            const infoTileTechniqueIndex = tileInfoWriter.addTechnique(technique);
-
-            let currentGeoJsonIndex = 0;
-            for (const aLine of lines) {
-                // add the geoJsonProperties for this line. undefined values are accepted as some
-                // line may not have any data.
-                const lineFeatureGroup = extendedTile.info.lineGroup as LineFeatureGroup;
-                if (lineFeatureGroup.userData === undefined) {
-                    lineFeatureGroup.userData = [geojsonProperties[currentGeoJsonIndex]];
-                } else {
-                    lineFeatureGroup.userData.push(geojsonProperties[currentGeoJsonIndex]);
-                }
-
-                currentGeoJsonIndex++;
-
-                tileInfoWriter.addFeature(
-                    extendedTile.info.lineGroup,
-                    technique,
-                    env,
-                    featureId,
-                    infoTileTechniqueIndex,
-                    false
-                );
-
-                tileInfoWriter.addFeaturePoints(extendedTile.info.lineGroup, aLine);
-            }
-
-            if (this.m_gatherRoadSegments) {
-                const segmentId = env.lookup("segmentId") as number;
-                if (segmentId !== undefined) {
-                    const startOffset = env.lookup("startOffset");
-                    const endOffset = env.lookup("endOffset");
-                    tileInfoWriter.addRoadSegments(
-                        extendedTile.info.lineGroup,
-                        segmentId,
-                        startOffset !== undefined ? (startOffset as number) : 0,
-                        endOffset !== undefined ? (endOffset as number) : 1
-                    );
-                }
-            }
-        }
-    }
-
-    private getTileInfo(extendedTile: ExtendedTile): ExtendedTileInfo {
-        extendedTile.writer.finish();
-        return extendedTile.info;
-    }
-
-    /**
-     * Processes GeoJSON "Polygon" and "MultiPolygon" objects.
-     *
-     * @param coordinates array containing line coordinates
-     * @param techniqueIndex technique index to render geometry fill
-     * @param geojsonProperties object containing the custom properties defined by the user
-     */
-    private processPolygon(
-        coordinates: number[][][][],
-        techniqueIndex: number,
-        geojsonProperties?: {}
-    ): void {
-        const buffer = this.findOrCreateGeometryBuffer(techniqueIndex);
-        buffer.type = "polygon";
-
-        coordinates.forEach(polygons => {
-            const vertices: number[] = [];
-            const holes: number[] = [];
-            polygons.forEach(polygon => {
-                if (polygon[0] === null || typeof polygon[0][0] !== "number") {
-                    return;
-                }
-                // the first polygon in the coordinates array is the main polygon the other ones
-                // are holes.
-                if (vertices.length) {
-                    holes.push(vertices.length / 2);
-                }
-
-                polygon.forEach(point => {
-                    this.m_cached_geoCoord.latitude = point[1];
-                    this.m_cached_geoCoord.longitude = point[0];
-
-                    this.m_projection
-                        .projectPoint(this.m_cached_geoCoord, this.m_cached_worldCoord)
-                        .sub(this.m_center);
-
-                    // polygon issue fix
-                    if (point[0] >= 180) {
-                        this.m_cached_worldCoord.x = this.m_cached_worldCoord.x * -1;
-                    }
-
-                    vertices.push(this.m_cached_worldCoord.x, this.m_cached_worldCoord.y);
-                });
-            });
-            buffer.polygons.push({ vertices, holes, geojsonProperties });
-        });
-    }
-
-    /**
-     * Creates a geometry from the previously filled in buffers.
-     */
-    private createGeometry(tileInfo: ExtendedTileInfo): void {
-        this.m_geometryBuffer.forEach((geometryData, technique) => {
-            switch (geometryData.type) {
-                case "point":
-                    this.createPointGeometry(technique, geometryData);
-                    break;
-                case "solid-line":
-                    this.createSolidLineGeometry(technique, geometryData);
-                    break;
-                case "polygon":
-                    this.createPolygonGeometry(technique, geometryData, tileInfo);
-                    break;
-                case "segments":
-                    this.createSegmentsGeometry(technique, geometryData);
-                    break;
-            }
-        });
-        this.m_poiGeometryBuffer.forEach((geometryData, technique) => {
-            this.createPoiGeometry(technique, geometryData);
-        });
-        this.m_textGeometryBuffer.forEach((geometryData, technique) => {
-            this.createTextGeometry(technique, geometryData);
-        });
-        this.m_textPathGeometryBuffer.forEach((geometryData, technique) => {
-            this.createTextPathGeometry(technique, geometryData);
-        });
-    }
-
-    private createPoiGeometry(technique: number, geometryData: GeometryData): void {
-        const geometry: GeoJsonPoiGeometry = {
-            positions: {
-                name: "position",
-                type: "float",
-                itemCount: 3,
-                buffer: new Float32Array(geometryData.points.vertices).buffer
-            },
-            technique,
-            texts: [0],
-            objInfos: geometryData.points.geojsonProperties
-        };
-        if (this.m_poiGeometries === undefined) {
-            this.m_poiGeometries = [];
-        }
-        this.m_poiGeometries.push(geometry);
-    }
-
-    private createTextGeometry(technique: number, geometryData: GeometryData): void {
-        const labelProperty = geometryData.labelProperty! as string;
-        const properties = geometryData.points.geojsonProperties[0];
-        const text = (properties as any)[labelProperty].toString();
-        const geometry: GeoJsonTextGeometry = {
-            positions: {
-                name: "position",
-                type: "float",
-                itemCount: 3,
-                buffer: new Float32Array(geometryData.points.vertices).buffer
-            },
-            technique,
-            texts: [0],
-            stringCatalog: [text],
-            objInfos: geometryData.points.geojsonProperties
-        };
-        if (this.m_textGeometries === undefined) {
-            this.m_textGeometries = [];
-        }
-        this.m_textGeometries.push(geometry);
-    }
-
-    private createTextPathGeometry(technique: number, geometryData: GeometryData): void {
-        const lines = geometryData.lines;
-        for (let i = 0; i < geometryData.lines.vertices.length; i++) {
-            const pathVertex = lines.vertices[i];
-            const path: number[] = [];
-            path.push(...pathVertex);
-            const pathLengthSqr = Math2D.computeSquaredLineLength(path);
-            const properties = geometryData.lines.geojsonProperties[i];
-            const text = (properties as any)[geometryData.labelProperty!].toString();
-            const geometry: GeoJsonTextPathGeometry = {
-                technique,
-                path,
-                pathLengthSqr,
-                text,
-                objInfos: geometryData.lines.geojsonProperties
-            };
-            if (this.m_textPathGeometries === undefined) {
-                this.m_textPathGeometries = [];
-            }
-            this.m_textPathGeometries.push(geometry);
-        }
-    }
-
-    /**
-     * Creates points geometry. All points with the same technique are batched as one geometry.
-     *
-     * @param technique Technique number.
-     * @param geometryData Geometry data.
-     */
-    private createPointGeometry(technique: number, geometryData: GeometryData): void {
-        const geometry: GeoJsonGeometry = {
-            type: GeometryType.Point,
-            vertexAttributes: [
-                {
-                    name: "position",
-                    buffer: new Float32Array(geometryData.points.vertices).buffer as ArrayBuffer,
-                    itemCount: 3,
-                    type: "float"
-                }
-            ],
-            groups: [
-                {
-                    start: 0,
-                    count: 0,
-                    technique
-                }
-            ]
-        };
-        geometry.objInfos = geometryData.points.geojsonProperties;
-        this.m_geometries.push(geometry);
-    }
-
-    /**
-     * Creates solid line geometry. All lines with the same technique are batched as one geometry.
-     *
-     * @param technique Technique number.
-     * @param geometryData Geometry data.
-     */
-    private createSolidLineGeometry(technique: number, geometryData: GeometryData): void {
-        const lines = new Lines();
-        const positions = new Array<number>();
-        geometryData.lines.vertices.forEach(linePositions => {
-            lines.add(linePositions);
-            positions.push.apply(positions, linePositions);
-        });
-
-        const { vertices, indices } = lines;
-
-        const buffer = new Float32Array(vertices).buffer as ArrayBuffer;
-        const index = new Uint32Array(indices).buffer as ArrayBuffer;
-        const attr: InterleavedBufferAttribute = {
-            type: "float",
-            stride: 12,
-            buffer,
-            attributes: LINE_VERTEX_ATTRIBUTE_DESCRIPTORS
-        };
-
-        const geometries: GeoJsonGeometry = {
-            type: GeometryType.SolidLine,
-            index: {
-                buffer: index,
-                itemCount: 1,
-                type: "uint32",
-                name: "index"
-            },
-            interleavedVertexAttributes: [attr],
-            vertexAttributes: [
-                {
-                    name: "points",
-                    buffer: new Float32Array(positions).buffer as ArrayBuffer,
-                    itemCount: 2,
-                    type: "float"
-                }
-            ],
-            groups: [
-                {
-                    start: 0,
-                    count: 0,
-                    technique
-                }
-            ]
-        };
-        this.m_geometries.push(geometries);
-    }
-
-    /**
-     * Creates polygon geometry. All polygons with the same technique are batched as one geometry.
-     *
-     * @param technique Technique number.
-     * @param geometryData Geometry data.
-     * @param tileInfo Information about the tiles
-     */
-    private createPolygonGeometry(
-        techniqueIndex: number,
-        geometryData: GeometryData,
-        tileInfo: ExtendedTileInfo
-    ): void {
-        const meshBuffer = new MeshBuffer();
-        const {
-            positions,
-            indices,
-            groups,
-            featureStarts,
-            featureIds,
-            geojsonProperties,
-            outlineIndices
-        } = meshBuffer;
-
-        let contour: number[];
-        const holesVertices: number[][] = [];
-        const technique = this.m_styleSetEvaluator.techniques[techniqueIndex];
-        const isFilled = isFillTechnique(technique);
-        const fillTechnique = technique as FillTechnique;
-        const outlineWidth = getPropertyValue(fillTechnique.lineWidth, tileInfo.tileKey.level);
-
-        const tileWorldBounds = new THREE.Box3();
-        const geoBox = webMercatorTilingScheme.getGeoBox(tileInfo.tileKey);
-        const tileBounds = this.m_projection.projectBox(geoBox, tileWorldBounds);
-        const center = new THREE.Vector3();
-        tileBounds.getCenter(center);
-        const tileWorldExtents = tileWorldBounds.max.sub(center).x;
-
-        geometryData.polygons.forEach(polygon => {
-            contour = polygon.holes.length
-                ? polygon.vertices.slice(0, polygon.holes[0] * 2)
-                : polygon.vertices;
-
-            const baseVertex = positions.length / 3;
-            //external ring
-            if (outlineWidth !== undefined && outlineWidth > 0.0 && isFilled) {
-                this.m_outliner.addEdges(baseVertex, contour, tileWorldExtents, outlineIndices);
-            }
-
-            // holes, if any
-            if (polygon.holes.length) {
-                for (let i = 0; i < polygon.holes.length; i++) {
-                    if (i === polygon.holes.length - 1) {
-                        holesVertices[i] = polygon.vertices.slice(polygon.holes[i] * 2);
-                    } else {
-                        holesVertices[i] = polygon.vertices.slice(
-                            polygon.holes[i] * 2,
-                            polygon.holes[i + 1] * 2
-                        );
-                    }
-
-                    if (outlineWidth !== undefined && outlineWidth > 0.0 && isFilled) {
-                        this.m_outliner.addEdges(
-                            polygon.holes[i],
-                            holesVertices[i],
-                            tileWorldExtents,
-                            outlineIndices
-                        );
-                    }
-                }
-            }
-
-            featureStarts.push(indices.length / 3);
-            // featureIds should be the id of the feature, but for geoJSON datasource we do not have
-            // it in integers, and we do not use them. Therefore, zeroes are added.
-            featureIds.push(0);
-            geojsonProperties.push(polygon.geojsonProperties);
-
-            for (let i = 0; i < polygon.vertices.length; i += 2) {
-                positions.push(polygon.vertices[i], polygon.vertices[i + 1], 0);
-            }
-
-            const triangles = earcut(polygon.vertices, polygon.holes);
-
-            for (let i = 0; i < triangles.length; i += 3) {
-                const v1 = triangles[i];
-                const v2 = triangles[i + 1];
-                const v3 = triangles[i + 2];
-                indices.push(v1 + baseVertex, v2 + baseVertex, v3 + baseVertex);
-            }
-        });
-        this.m_outliner.fill(outlineIndices);
-
-        if (indices.length > 0) {
-            groups.push({
-                start: 0,
-                count: indices.length,
-                technique: techniqueIndex
-            });
-        }
-
-        const geometry: GeoJsonGeometry = {
-            type: GeometryType.Polygon,
-            vertexAttributes: [
-                {
-                    name: "position",
-                    buffer: new Float32Array(meshBuffer.positions).buffer as ArrayBuffer,
-                    itemCount: 3,
-                    type: "float"
-                }
-            ],
-            groups: meshBuffer.groups
-        };
-
-        if (meshBuffer.indices.length > 0) {
-            geometry.index = {
-                name: "index",
-                buffer: new Uint32Array(meshBuffer.indices).buffer as ArrayBuffer,
-                itemCount: 1,
-                type: "uint32"
-            };
-            geometry.featureStarts = meshBuffer.featureStarts;
-            geometry.featureIds = meshBuffer.featureIds;
-            geometry.objInfos = meshBuffer.geojsonProperties;
-        }
-
-        this.m_outliner.addOutlineIndicesArrayToGeometry(meshBuffer, geometry);
-        this.m_geometries.push(geometry);
-    }
-
-    /**
-     * Creates segments geometry used for polygon borders and invisible lines for raycasting. All
-     * segment lines with the same technique created as one geometry.
-     *
-     * @param technique Technique number.
-     * @param geometryData Geometry data.
-     */
-    private createSegmentsGeometry(technique: number, geometryData: GeometryData): void {
-        const geometries: GeoJsonGeometry = {
-            type: GeometryType.Line,
-            vertexAttributes: [
-                {
-                    name: "position",
-                    buffer: new Float32Array(geometryData.points.vertices).buffer as ArrayBuffer,
-                    itemCount: 3,
-                    type: "float"
-                }
-            ],
-            groups: [
-                {
-                    start: 0,
-                    count: 0,
-                    technique
-                }
-            ]
-        };
-        this.m_geometries.push(geometries);
-    }
-
-    /**
-     * Returns the indices of the techniques matching the feature type and feature properties.
-     *
-     * @param feature GeoJSON feature.
-     * @param envType Technique type name.
-     */
-    private findTechniqueIndices(feature: Feature, envType: Value): number[] {
-        const featureDetails: FeatureDetails = Flattener.flatten(feature.properties, "properties");
-        featureDetails.featureId = feature.id;
-        const env = new MapEnv({ type: envType, ...featureDetails });
-        return this.m_styleSetEvaluator.getMatchingTechniques(env).map(technique => {
-            const index = technique._index;
-            if (index === undefined) {
-                throw new Error(
-                    "GeoJsonDecoder#findOrCreateTechniqueIndex: Internal error - No technique index"
-                );
-            }
-            return index;
-        });
-    }
-
-    /**
-     * Creates a geometry buffer for the specified technique, or returns the existing one.
-     *
-     * @param index technique index
-     */
-    private findOrCreateGeometryBuffer(index: number): GeometryData {
-        let buffer = this.m_geometryBuffer.get(index);
-        if (buffer !== undefined) {
-            return buffer;
-        }
-        buffer = new GeometryData();
-        this.m_geometryBuffer.set(index, buffer);
-        return buffer;
-    }
-
-    /**
-     * Creates a POI geometry buffer for the specified technique, or returns the existing one.
-     *
-     * @param index technique index
-     */
-    private findOrCreatePoiGeometryBuffer(index: number): GeometryData {
-        let buffer = this.m_poiGeometryBuffer.get(index);
-        if (buffer !== undefined) {
-            return buffer;
-        }
-        buffer = new GeometryData();
-        this.m_poiGeometryBuffer.set(index, buffer);
-        return buffer;
-    }
-
-    /**
-     * Creates a Text geometry buffer for the specified technique, or returns the existing one.
-     *
-     * @param index technique index
-     */
-    private findOrCreateTextGeometryBuffer(index: number): GeometryData {
-        let buffer = this.m_textGeometryBuffer.get(index);
-        if (buffer !== undefined) {
-            return buffer;
-        }
-        buffer = new GeometryData();
-        this.m_textGeometryBuffer.set(index, buffer);
-        return buffer;
-    }
-
-    /**
-     * Creates a TextPath geometry buffer for the specified technique, or returns the existing one.
-     *
-     * @param index technique index
-     */
-    private findOrCreateTextPathGeometryBuffer(index: number): GeometryData {
-        let buffer = this.m_textPathGeometryBuffer.get(index);
-        if (buffer !== undefined) {
-            return buffer;
-        }
-        buffer = new GeometryData();
-        this.m_textPathGeometryBuffer.set(index, buffer);
-        return buffer;
     }
 
     private getTileCenter(tileKey: TileKey) {
@@ -1250,6 +154,11 @@ class GeoJsonDecoder {
         const center = new THREE.Vector3();
         tileBounds.getCenter(center);
         return center;
+    }
+
+    private getTileInfo(extendedTile: ExtendedTile): ExtendedTileInfo {
+        extendedTile.writer.finish();
+        return extendedTile.info;
     }
 }
 
@@ -1271,7 +180,7 @@ export class GeoJsonTileDecoderService {
 }
 
 /**
- * Starts an OMV decoder service.
+ * Starts a GeoJson decoder service.
  *
  * @deprecated Please use [[GeoJsonTileDecoderService.start]].
  */
