+++ @here/harp-mapview/lib/Tile.ts
@@ -1057,7 +1057,7 @@ export class Tile implements CachedResource {
 
                 if (material === undefined) {
                     material = createMaterial({
-                        technique: decodedTile.techniques[techniqueIndex],
+                        technique,
                         level: displayZoomLevel,
                         fog: this.mapView.scene.fog !== null
                     });
@@ -1324,7 +1324,6 @@ export class Tile implements CachedResource {
                         // Configure the edge material based on the theme values.
                         const materialParams: EdgeMaterialParameters = {
                             color: fadingParams.color,
-                            colorMix: fadingParams.colorMix,
                             fadeNear: fadingParams.lineFadeNear,
                             fadeFar: fadingParams.lineFadeFar
                         };
@@ -1360,19 +1359,21 @@ export class Tile implements CachedResource {
      * Counts the number of vertices in a tile.
      */
     countVertices(): void {
-        this.objects.filter(object => object instanceof THREE.Mesh).forEach(object => {
-            const mesh = object as THREE.Mesh;
-            if (mesh.geometry instanceof THREE.BufferGeometry) {
-                if (mesh.geometry.index !== undefined && mesh.geometry.index !== null) {
-                    this.m_numVertices += mesh.geometry.index.count;
-                } else {
-                    this.m_numVertices += mesh.geometry.getAttribute("position").count / 3;
+        this.objects
+            .filter(object => object instanceof THREE.Mesh)
+            .forEach(object => {
+                const mesh = object as THREE.Mesh;
+                if (mesh.geometry instanceof THREE.BufferGeometry) {
+                    if (mesh.geometry.index !== undefined && mesh.geometry.index !== null) {
+                        this.m_numVertices += mesh.geometry.index.count;
+                    } else {
+                        this.m_numVertices += mesh.geometry.getAttribute("position").count / 3;
+                    }
                 }
-            }
-            if (mesh.geometry instanceof THREE.Geometry) {
-                this.m_numVertices = mesh.geometry.vertices.length;
-            }
-        });
+                if (mesh.geometry instanceof THREE.Geometry) {
+                    this.m_numVertices = mesh.geometry.vertices.length;
+                }
+            });
     }
 
     /**
